server/requirements.txt
flask
web3
eth_account
python-dotenv
flask_cors
psycopg2
markdown